requirements.txt
flask>=3.0,<4.0
Pillow>=10.0
